+++ packages/core-js/actual/iterator/zip-keyed.js
@@ -2,6 +2,7 @@
 var parent = require('../../stable/iterator/zip');
 require('../../modules/esnext.iterator.chunks');
 require('../../modules/esnext.iterator.includes');
+require('../../modules/esnext.iterator.join');
 require('../../modules/esnext.iterator.windows');
 require('../../modules/esnext.iterator.zip-keyed');