+++ packages/core-js/full/iterator/range.js
@@ -18,6 +18,7 @@ require('../../modules/es.iterator.to-array');
 require('../../modules/esnext.iterator.constructor');
 require('../../modules/esnext.iterator.chunks');
 require('../../modules/esnext.iterator.includes');
+require('../../modules/esnext.iterator.join');
 require('../../modules/esnext.iterator.range');
 require('../../modules/esnext.iterator.sliding');
 require('../../modules/esnext.iterator.windows');