+++ packages/core-js/stage/3.js
@@ -3,6 +3,7 @@ var parent = require('./4');
 
 require('../proposals/decorator-metadata-v2');
 require('../proposals/iterator-includes');
+require('../proposals/iterator-join');
 require('../proposals/joint-iteration');
 // TODO: Obsolete versions, remove from `core-js@4`
 require('../proposals/array-grouping-stage-3');