+++ tests/compat/tests.js
@@ -1958,6 +1958,9 @@ GLOBAL.tests = {
   'esnext.iterator.includes': function () {
     return Iterator.prototype.includes;
   },
+  'esnext.iterator.join': function () {
+    return Iterator.prototype.join;
+  },
   'esnext.iterator.range': function () {
     return Iterator.range;
   },