+++ tests/eslint/eslint.config.js
@@ -1301,6 +1301,7 @@ const forbidCompletelyNonExistentBuiltIns = {
   'es/no-nonstandard-iterator-prototype-properties': [ERROR, { allow: [
     'chunks',
     'includes',
+    'join',
     'sliding',
     'toAsync',
     'windows',